+++ common/src/main/java/net/arna/jcraft/common/util/IClientEntityHandler.java
@@ -17,4 +17,6 @@ public interface IClientEntityHandler {
     void purpleHazeRemoteClientTick(AbstractPurpleHazeEntity<?, ?> purpleHazeEntity);
 
     void sheerHeartAttackEntityTick(SheerHeartAttackEntity sHAEntity);
+
+    void displayMetallicaAura(MetallicaEntity metallica);
 }